requirements.txt
numpy==1.26.4
torch==2.1.0
torchvision==0.16.0
monai==1.3.0
SimpleITK
nibabel
pydicom
scikit-learn
scipy
pandas
matplotlib
seaborn
plotly
gradio
fastapi
uvicorn
wandb
captum
connected-components-3d
gdown
tqdm
pyyaml
omegaconf
pycox